backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
distlib==0.3.9
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
filelock==3.18.0
packaging==24.2
pipenv==2024.4.1
platformdirs==4.3.7
setuptools==78.1.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.30.0